# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "AppearanceInput": {
        "additionalProperties": false,
        "properties": {
          "captain": {
            "default": false,
            "title": "Captain",
            "type": "boolean"
          },
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "position_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Id"
          },
          "shirt_number": {
            "anyOf": [
              {
                "maximum": 99.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shirt Number"
          },
          "started": {
            "default": true,
            "title": "Started",
            "type": "boolean"
          }
        },
        "required": [
          "player_id"
        ],
        "title": "AppearanceInput",
        "type": "object"
      },
      "AppearanceRead": {
        "properties": {
          "captain": {
            "title": "Captain",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerSummary"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "shirt_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shirt Number"
          },
          "started": {
            "title": "Started",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "player",
          "started",
          "position",
          "shirt_number",
          "captain"
        ],
        "title": "AppearanceRead",
        "type": "object"
      },
...
      "AwardInput": {
        "additionalProperties": false,
        "properties": {
          "award_type_id": {
            "title": "Award Type Id",
            "type": "integer"
          },
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          }
        },
        "required": [
          "award_type_id",
          "player_id"
        ],
        "title": "AwardInput",
        "type": "object"
      },
      "AwardRead": {
        "properties": {
          "award_type": {
            "$ref": "#/components/schemas/AwardTypeRead"
          },
          "fixture_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixture Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "period_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Label"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerSummary"
          }
        },
        "required": [
          "id",
          "award_type",
          "player",
          "fixture_id",
          "period_label",
          "notes"
        ],
        "title": "AwardRead",
        "type": "object"
      },
...
      "CompetitionRead": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/CompetitionType"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "is_active"
        ],
        "title": "CompetitionRead",
        "type": "object"
      },
      "CompetitionType": {
        "enum": [
          "league",
          "cup",
          "friendly",
          "tournament"
        ],
        "title": "CompetitionType",
        "type": "string"
      },
...
      "FixtureCreate": {
        "additionalProperties": false,
        "properties": {
          "competition_id": {
            "title": "Competition Id",
            "type": "integer"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 10.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "kickoff_at": {
            "format": "date-time",
            "title": "Kickoff At",
            "type": "string"
          },
          "match_number": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Number"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "opposition_team_id": {
            "title": "Opposition Team Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/FixtureStatus",
            "default": "scheduled"
          },
          "team_season_id": {
            "title": "Team Season Id",
            "type": "integer"
          },
          "venue": {
            "$ref": "#/components/schemas/Venue",
            "default": "home"
          },
          "venue_notes": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Notes"
          }
        },
        "required": [
          "team_season_id",
          "competition_id",
          "opposition_team_id",
          "kickoff_at"
        ],
        "title": "FixtureCreate",
        "type": "object"
      },
      "FixtureDetail": {
        "properties": {
          "appearances": {
            "items": {
              "$ref": "#/components/schemas/AppearanceRead"
            },
            "title": "Appearances",
            "type": "array"
          },
          "awards": {
            "items": {
              "$ref": "#/components/schemas/AwardRead"
            },
            "title": "Awards",
            "type": "array"
          },
          "competition": {
            "$ref": "#/components/schemas/CompetitionRead"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/GoalRead"
            },
            "title": "Goals",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kickoff_at": {
            "format": "date-time",
            "title": "Kickoff At",
            "type": "string"
          },
          "match_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Number"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "opposition": {
            "$ref": "#/components/schemas/TeamRead"
          },
          "our_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Our Score"
          },
          "status": {
            "$ref": "#/components/schemas/FixtureStatus"
          },
          "team_season_id": {
            "title": "Team Season Id",
            "type": "integer"
          },
          "their_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Their Score"
          },
          "venue": {
            "$ref": "#/components/schemas/Venue"
          },
          "venue_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Notes"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "team_season_id",
          "competition",
          "opposition",
          "match_number",
          "kickoff_at",
          "venue",
          "venue_notes",
          "status",
          "our_score",
          "their_score",
          "duration_minutes",
          "notes",
          "appearances",
          "goals",
          "awards",
          "warnings"
        ],
        "title": "FixtureDetail",
        "type": "object"
      },
      "FixtureRead": {
        "properties": {
          "competition": {
            "$ref": "#/components/schemas/CompetitionRead"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kickoff_at": {
            "format": "date-time",
            "title": "Kickoff At",
            "type": "string"
          },
          "match_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Number"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "opposition": {
            "$ref": "#/components/schemas/TeamRead"
          },
          "our_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Our Score"
          },
          "status": {
            "$ref": "#/components/schemas/FixtureStatus"
          },
          "team_season_id": {
            "title": "Team Season Id",
            "type": "integer"
          },
          "their_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Their Score"
          },
          "venue": {
            "$ref": "#/components/schemas/Venue"
          },
          "venue_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Notes"
          }
        },
        "required": [
          "id",
          "team_season_id",
          "competition",
          "opposition",
          "match_number",
          "kickoff_at",
          "venue",
          "venue_notes",
          "status",
          "our_score",
          "their_score",
          "duration_minutes",
          "notes"
        ],
        "title": "FixtureRead",
        "type": "object"
      },
      "FixtureStatus": {
        "enum": [
          "scheduled",
          "played",
          "postponed",
          "cancelled",
          "abandoned"
        ],
        "title": "FixtureStatus",
        "type": "string"
      },
      "FixtureUpdate": {
        "additionalProperties": false,
        "properties": {
          "competition_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competition Id"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 10.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "kickoff_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kickoff At"
          },
          "match_number": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Number"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "opposition_team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opposition Team Id"
          },
          "our_score": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Our Score"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FixtureStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "their_score": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Their Score"
          },
          "venue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Venue"
              },
              {
                "type": "null"
              }
            ]
          },
          "venue_notes": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Notes"
          }
        },
        "title": "FixtureUpdate",
        "type": "object"
      },
...
      "GoalInput": {
        "additionalProperties": false,
        "description": "scorer_id None + event_type opp_own_goal = opposition own goal.",
        "properties": {
          "assisted_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assisted By Id"
          },
          "event_type": {
            "default": "goal",
            "pattern": "^(goal|own_goal|opp_own_goal)$",
            "title": "Event Type",
            "type": "string"
          },
          "minute": {
            "anyOf": [
              {
                "maximum": 130.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "scorer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scorer Id"
          }
        },
        "title": "GoalInput",
        "type": "object"
      },
      "GoalRead": {
        "description": "A goal event with its assist folded in - the shape the UI wants.",
        "properties": {
          "assist_event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assist Event Id"
          },
          "assisted_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "scorer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "event_type",
          "scorer",
          "assisted_by",
          "assist_event_id",
          "minute",
          "sequence",
          "notes"
        ],
        "title": "GoalRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlayerSummary": {
        "description": "Lightweight embed for nested responses.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "PlayerSummary",
        "type": "object"
      },
...
      "ResultSubmit": {
        "additionalProperties": false,
        "description": "The whole result of a match in one transactional write.",
        "properties": {
          "appearances": {
            "items": {
              "$ref": "#/components/schemas/AppearanceInput"
            },
            "title": "Appearances",
            "type": "array"
          },
          "awards": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AwardInput"
            },
            "title": "Awards",
            "type": "array"
          },
          "goals": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GoalInput"
            },
            "title": "Goals",
            "type": "array"
          },
          "our_score": {
            "minimum": 0.0,
            "title": "Our Score",
            "type": "integer"
          },
          "their_score": {
            "minimum": 0.0,
            "title": "Their Score",
            "type": "integer"
          }
        },
        "required": [
          "our_score",
          "their_score",
          "appearances"
        ],
        "title": "ResultSubmit",
        "type": "object"
      },
...
      "TeamRead": {
        "properties": {
          "colours": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colours"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          }
        },
        "required": [
          "id",
          "name",
          "short_name",
          "colours",
          "notes"
        ],
        "title": "TeamRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Venue": {
        "enum": [
          "home",
          "away",
          "neutral"
        ],
        "title": "Venue",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "ABGFC Blues",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/fixtures": {
      "get": {
        "operationId": "list_fixtures_api_v1_fixtures_get",
        "parameters": [
          {
            "in": "query",
            "name": "team_season_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Season Id"
            }
          },
          {
            "in": "query",
            "name": "competition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Competition Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FixtureStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FixtureRead"
                  },
                  "title": "Response List Fixtures Api V1 Fixtures Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Fixtures",
        "tags": [
          "fixtures"
        ]
      },
      "post": {
        "operationId": "create_fixture_api_v1_fixtures_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixtureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Fixture",
        "tags": [
          "fixtures"
        ]
      }
    },
    "/api/v1/fixtures/{fixture_id}": {
      "delete": {
        "operationId": "delete_fixture_api_v1_fixtures__fixture_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Fixture",
        "tags": [
          "fixtures"
        ]
      },
      "get": {
        "operationId": "get_fixture_api_v1_fixtures__fixture_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Fixture",
        "tags": [
          "fixtures"
        ]
      },
      "patch": {
        "operationId": "update_fixture_api_v1_fixtures__fixture_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixtureUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Fixture",
        "tags": [
          "fixtures"
        ]
      }
    },
    "/api/v1/fixtures/{fixture_id}/result": {
      "put": {
        "description": "Post-match entry: appearances, goals/assists, awards and score in one write.",
        "operationId": "submit_result_api_v1_fixtures__fixture_id__result_put",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Result",
        "tags": [
          "fixtures"
        ]
      }
    },
...
  }
}